fix(i18n): SSR locale sync and desktop language toggle
Run locale resolution on server+client so waro_locale=en does not flash
Spanish. Add es/en controls to the desktop tenant/settings modal (mobile
already had them). Keep cookie path=/ for consistent reloads.

Refs #1602
