+++ docs/engineering/i18n.md
@@ -40,6 +40,7 @@ Shell chrome only:
 ## Manual smoke
 
 1. Cold load with no cookie → Spanish shell matches production copy.  
-2. Settings → Language → English → nav/auth/header shell strings switch; module body stays Spanish.  
-3. Reload → cookie keeps English.  
-4. Clear cookie → back to `es` (or tenant locale when B1 is live).
+2. Language toggle: **desktop** account/tenant modal or **mobile** Settings → Language → English → nav/auth/header shell strings switch; module body stays Spanish.  
+3. Reload with `waro_locale=en` → first HTML should already be English (SSR plugin, not client-only).  
+4. Clear cookie → back to `es` (or tenant locale when B1 is live).  
+5. Logout clears `localStorage` but **not** the `waro_locale` cookie.
